Renamed springuni text to craftingjava in all files.
* Renamed springuni-examples to craftingjava-examples
* Renamed packages from com.springuni to com.craftingjava
that
* Replaced all occurrences of "SpringUni" to "CraftingJava" if you find it in any text files (including source)
