+++ pom.xml
@@ -4,10 +4,10 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.springuni.examples</groupId>
-  <artifactId>springuni-examples</artifactId>
+  <groupId>com.craftingjava.examples</groupId>
+  <artifactId>craftingjava-examples</artifactId>
   <version>1.0.0-SNAPSHOT</version>
-  <name>SpringUni Examples</name>
+  <name>CraftingJava Examples</name>
 
   <modules>
     <module>spring-boot</module>
@@ -19,7 +19,7 @@
 
   <organization>
     <name>Laszlo Csontos (sole trader)</name>
-    <url>https://springuni.com/</url>
+    <url>https://craftingjava.com/</url>
   </organization>
 
   <licenses>
@@ -30,8 +30,8 @@
   </licenses>
 
   <scm>
-    <connection>scm:git:git@github.com:springuni/springuni-examples.git</connection>
-    <url>https://github.com/springuni/springuni-examples</url>
+    <connection>scm:git:git@github.com:craftingjava/craftingjava-examples.git</connection>
+    <url>https://github.com/craftingjava/craftingjava-examples</url>
     <tag>HEAD</tag>
   </scm>
