+++ spring-boot/spring-jms-sqs/src/main/java/com/craftingjava/examples/jms/sqs/StandardSqsConfiguration.java
@@ -1,0 +1,19 @@
+package com.craftingjava.examples.jms.sqs;
+
+import javax.jms.ConnectionFactory;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Created by lcsontos on 8/4/17.
+ */
+@Configuration
+public class StandardSqsConfiguration extends AbstractSqsConfiguration {
+
+  @Bean
+  @Override
+  public ConnectionFactory connectionFactory(SqsProperties sqsProperties) {
+    return createStandardSQSConnectionFactory(sqsProperties);
+  }
+
+}
