--- spring-boot/spring-jms-sqs/src/main/java/com/springuni/examples/jms/sqs/StandardSqsConfiguration.java
@@ -1,19 +1,0 @@
-package com.springuni.examples.jms.sqs;
-
-import javax.jms.ConnectionFactory;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * Created by lcsontos on 8/4/17.
- */
-@Configuration
-public class StandardSqsConfiguration extends AbstractSqsConfiguration {
-
-  @Bean
-  @Override
-  public ConnectionFactory connectionFactory(SqsProperties sqsProperties) {
-    return createStandardSQSConnectionFactory(sqsProperties);
-  }
-
-}
