+++ spring-cloud/pom.xml
@@ -5,8 +5,8 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>com.springuni.examples</groupId>
-    <artifactId>springuni-examples</artifactId>
+    <groupId>com.craftingjava.examples</groupId>
+    <artifactId>craftingjava-examples</artifactId>
     <version>1.0.0-SNAPSHOT</version>
   </parent>
