+++ spring-cloud/spring-cloud-kubernetes/kubernetes-configmap/src/main/java/com/craftingjava/examples/kubernetes/KubernetesConfigMapApplication.java
@@ -1,0 +1,13 @@
+package com.craftingjava.examples.kubernetes;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class KubernetesConfigMapApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(KubernetesConfigMapApplication.class, args);
+  }
+
+}
