--- spring-cloud/spring-cloud-kubernetes/kubernetes-configmap/src/main/java/com/springuni/examples/kubernetes/KubernetesConfigMapApplication.java
@@ -1,13 +1,0 @@
-package com.springuni.examples.kubernetes;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class KubernetesConfigMapApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(KubernetesConfigMapApplication.class, args);
-  }
-
-}
