+++ spring-cloud/spring-cloud-kubernetes/kubernetes-core/src/main/java/com/craftingjava/examples/kubernetes/KubernetesCoreApplication.java
@@ -1,0 +1,15 @@
+package com.craftingjava.examples.kubernetes;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class KubernetesCoreApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(KubernetesCoreApplication.class, args);
+  }
+
+}
