+++ spring-cloud/spring-cloud-kubernetes/pom.xml
@@ -5,7 +5,7 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>com.springuni.examples</groupId>
+    <groupId>com.craftingjava.examples</groupId>
     <artifactId>spring-cloud</artifactId>
     <version>1.0.0-SNAPSHOT</version>
   </parent>
