+++ spring-cloud/spring-cloud-stream-pubsub-publisher/src/main/java/com/craftingjava/examples/pubsub/publisher/Channels.java
@@ -1,0 +1,11 @@
+package com.craftingjava.examples.pubsub.publisher;
+
+import org.springframework.cloud.stream.annotation.Output;
+import org.springframework.messaging.MessageChannel;
+
+public interface Channels {
+
+  @Output
+  MessageChannel outgoing();
+
+}
