--- spring-cloud/spring-cloud-stream-pubsub-publisher/src/main/java/com/springuni/examples/pubsub/publisher/Channels.java
@@ -1,11 +1,0 @@
-package com.springuni.examples.pubsub.publisher;
-
-import org.springframework.cloud.stream.annotation.Output;
-import org.springframework.messaging.MessageChannel;
-
-public interface Channels {
-
-  @Output
-  MessageChannel outgoing();
-
-}
