--- spring-cloud/spring-cloud-stream-pubsub-publisher/src/main/java/com/springuni/examples/pubsub/publisher/PubsubPublisherApplication.java
@@ -1,15 +1,0 @@
-package com.springuni.examples.pubsub.publisher;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.stream.annotation.EnableBinding;
-
-@EnableBinding(Channels.class)
-@SpringBootApplication
-public class PubsubPublisherApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(PubsubPublisherApplication.class, args);
-  }
-
-}
