+++ spring-cloud/spring-cloud-stream-pubsub-publisher/src/test/java/com/craftingjava/examples/pubsub/publisher/PubsubPublisherApplicationIT.java
@@ -1,0 +1,16 @@
+package com.craftingjava.examples.pubsub.publisher;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class PubsubPublisherApplicationIT {
+
+  @Test
+  public void contextLoads() {
+  }
+
+}
