--- spring-cloud/spring-cloud-stream-pubsub-publisher/src/test/java/com/springuni/examples/pubsub/publisher/PubsubPublisherApplicationTests.java
@@ -1,16 +1,0 @@
-package com.springuni.examples.pubsub.publisher;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class PubsubPublisherApplicationTests {
-
-  @Test
-  public void contextLoads() {
-  }
-
-}
