+++ spring-cloud/spring-cloud-stream-pubsub-subscriber/pom.xml
@@ -4,7 +4,7 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.springuni.examples</groupId>
+  <groupId>com.craftingjava.examples</groupId>
   <artifactId>spring-cloud-stream-pubsub-subscriber</artifactId>
   <version>0.0.1-SNAPSHOT</version>
   <packaging>jar</packaging>
