+++ spring-cloud/spring-cloud-stream-pubsub-subscriber/src/test/java/com/craftingjava/examples/pubsub/subscriber/PubsubSubscriberApplicationIT.java
@@ -1,0 +1,16 @@
+package com.craftingjava.examples.pubsub.subscriber;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class PubsubSubscriberApplicationIT {
+
+  @Test
+  public void contextLoads() {
+  }
+
+}
