--- spring-cloud/spring-cloud-stream-pubsub-subscriber/src/test/java/com/springuni/examples/pubsub/subscriber/PubsubSubscriberApplicationTests.java
@@ -1,16 +1,0 @@
-package com.springuni.examples.pubsub.subscriber;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class PubsubSubscriberApplicationTests {
-
-  @Test
-  public void contextLoads() {
-  }
-
-}
