+++ spring-integration/twitter-streaming/src/main/java/com/craftingjava/examples/twitter/TwitterStreamApplication.java
@@ -1,0 +1,13 @@
+package com.craftingjava.examples.twitter;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TwitterStreamApplication {
+
+  public static void main(String[] args) throws Exception {
+    SpringApplication.run(TwitterStreamApplication.class, args);
+  }
+
+}
