--- spring-integration/twitter-streaming/src/main/java/com/springuni/examples/twitter/TwitterStreamApplication.java
@@ -1,13 +1,0 @@
-package com.springuni.examples.twitter;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TwitterStreamApplication {
-
-  public static void main(String[] args) throws Exception {
-    SpringApplication.run(TwitterStreamApplication.class, args);
-  }
-
-}
